fix dot tests; update runtime test funcs
