+++ dot/core/service.go
@@ -643,7 +643,7 @@ func (s *Service) checkForRuntimeChanges() error {
 
 		s.rt.Stop()
 
-		s.rt, err = runtime.NewRuntime(code, s.storageState, s.keys)
+		s.rt, err = runtime.NewRuntime(code, s.storageState, s.keys, runtime.RegisterImportsOld)
 		if err != nil {
 			return err
 		}
