+++ dot/genesis_test.go
@@ -34,7 +34,6 @@ import (
 	"github.com/ChainSafe/gossamer/lib/runtime"
 	"github.com/ChainSafe/gossamer/lib/trie"
 	"github.com/ChainSafe/gossamer/lib/utils"
-	"github.com/ChainSafe/gossamer/tests"
 
 	"github.com/stretchr/testify/require"
 )
@@ -44,8 +43,8 @@ import (
 func newTestGenesisAndRuntime(t *testing.T) string {
 	dir := utils.NewTestDir(t)
 
-	_ = runtime.NewTestRuntime(t, tests.POLKADOT_RUNTIME)
-	runtimeFilePath := tests.GetAbsolutePath(tests.POLKADOT_RUNTIME_FP)
+	_ = runtime.NewTestRuntime(t, runtime.POLKADOT_RUNTIME)
+	runtimeFilePath := runtime.GetAbsolutePath(runtime.POLKADOT_RUNTIME_FP)
 
 	runtimeData, err := ioutil.ReadFile(runtimeFilePath)
 	require.Nil(t, err)
